requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
numpy<2.0.0
pandas
openpyxl
insightface
onnxruntime
opencv-python-headless
jinja2
sqlalchemy
asyncpg
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]